[MSFT] Making `PassCommon` public
Other dialects could make use of this functionality.
